[BTD-541] Added check match end point
